+++ packages/math-input/src/components/input/mathquill-instance.ts
@@ -1,0 +1,5 @@
+import MathQuill from "mathquill";
+
+import {MathQuillInterface} from "./mathquill-types";
+
+export default MathQuill.getInterface(2) as MathQuillInterface;