+++ packages/perseus/src/widgets/explanation.tsx
@@ -13,7 +13,7 @@ import type {PerseusExplanationWidgetOptions} from "../perseus-types";
 import type {PerseusScore, WidgetExports, WidgetProps} from "../types";
 
 // TODO(FEI-5054): Figure out how to get global .d.ts files working with monorepos
-type Empty = Record<string, never>;
+type Empty = Record<never, never>;
 
 type RenderProps = PerseusExplanationWidgetOptions; // transform = _.identity