+++ packages/perseus/src/widgets/image.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable react/prop-types */
 /* eslint-disable react/sort-comp */
 import {linterContextDefault} from "@khanacademy/perseus-linter";
 import {StyleSheet, css} from "aphrodite";